Show parameters of transition actions
One can do several actions during a transition in the process management.
For each action a TransitionAction module exists and those actions
need various parameters.

As it is hard to remember all the parameters, those parameters should be
shown in the dialog when an action module is selected. This commit implements
this behaviour.

Nevertheless, the reference that is shown under the parameter list could be
nicer. But I'm not a CSS ninja ;-)