Issue #2251: use the special cases in SUPER::ValueIsDifferent()
The method Kernel::System::DynamicField::Driver::Base::ValueIsDifferent()
the handled special cases concerned with empty strings and undef.
Implementing ValueIsDifferent() in a derived class effectively removed
these special cases. Let's add them back in, so that TicketHistoryDynamicField.t
succeeds again.
Note that this is not the final solution, as there are more cases to be
considered.