Issue #2298: pass dynamic field values in _Mask() helpers
A recent addition is that helper subs like _Mask() and _MaskPhone() now get
the additional parameter DFValues. But these helper subs are also called
in the case of error, or when a form draft was loaded. In these instances
the new parameter DFValues was missing.
This patch should fix three test failures.