Issue #2711: add a sample file for an unitest
The relavant unit test is scripts/test/Selenium/Agent/Admin/AdminImportExport.t
from the ITSMConfigurationManagement package.
The sample is included in the OTOBO core repository because it must
be available in the Selenium Docker image. The Selenium image is built
based on the core repository and has no connection to the
ITSMConfigurationManagement package.