Issue #3249: Start conversion of Safety() to HTML::Scrubber
Provide a local copy of HTML::Scrubber in Kernel/cpan-lib.
The module is in the OTOBO repository because we will enhance the module later on.
The local copy is taken from the HTML::Scrubber github repository.

The version is set to '019-OTOBO' in order to indicate OTOBO specific adaptions.
Let's see whether this version declaration is accepted.