Issue #1884: new icon for Wikipedia in customer interface
Downloaded a new W icon from iconfinder.com as the previous
hotlinked icon did no longer exist on the iconfinder CDN.
Use the stylized W instead of the puzzleglobe, as the puzzleglobe seems
to be the trademark of Wikipedia.
Add the icon as SVG data-url because this way allows to add icons
in the SysConfig without uploading any file.