Issue #3249: Let HTML::Parser remove specific tags.
When 'NoObject' is set then ignore both 'object' and 'param' tags.
As a side effect, 'Replace' is not set in these cases.

Also note, that the elements between the start and end tag should be removed.
But this is only the case when the nesting of the tags is sane.

Note that tags like <img/src="nice.jpg"> are not removed when only NoImg is passed.
This is because the tag name is recognized as 'img/src="nice.jpg"'.