Issue #3249: Check for protocol-relative URLs
when deciding whether URLs should be rewritten in order to include the session ID.
This is important because the session ID is leaked when it is appended
to external URLs.

Also make the check for external URLs less specific,
e.g. check for 'https:' instead of 'http:/'.

Also, Trim some whitespace where it is appropriate. And accept more whitespace
when looking at HTML with regexes. The URL in 'src' and 'href' attributes can indeed
be legally surrounded by leading and trailing whitespace. Text::Trim is not used
for trimming, as Text::Trim is not required for OTOBO 10.0.x

Check for protocol relevant links also in Redirect() even it that case
should not occur, as the code is called only for internal links.