Issue #3100: avoid timeouts in test execution
Checking for the expected output in WaitFor() yields a timeout
when there is a mismatch and not a helpful message.
Checking for a single line in the content yields a timeout
when the previously set data has more than a single line.

Also adapt the expected output for simple text. CKEditor 5 seems
to wrap that text into a <p> element.