Issue #3326: avoid assignment to the topic by diamond operator
In specific context, primarily in while conditions, the diamond operator
makes an assignment to the topic variable $_. Using this feature is not wrong.
However it is not good practice. $_ is a global variable and messing with globals
could lead to unwanted side effects.

Therefore honor the Perl::Critic policy Community::WhileDiamondDefaultAssignment
The approach is to make simple changes achieving the goal. There is
no migration to file slurping or Capture::Tiny.