Issue #3100: avoid high load when the LCSS is computed
In TestingMethods.t there are tests whether a string is a substring
that are expected to fail. These tests do fail indeed. Test::LongString then
tries to show the longest common substring, aka LCSS. For some reason the
reporting method produces high load and runs for a long time. That looks
like a bug in Test::LongString.

So, turn off LCSS reporting in order to avoid the high load.