Issue #3769: fix and enhance profiling
There was an error when profiling was deactivated, it propably had to do
with the DSL syntax of Plack::Builder.
So let's enable the route /otobo-web/nytprof in all cases.

For convenience create a new report dir for every request. This offers
the chance to see all requests, including the AJAX requests. Made a
workaround to make the profile id available in the report dirname.

Enhance the profile id to show the human readable date and time.