Issue #3937: update expected output for Cpanel::JSON::XS
Cpanel::JSON::XS is better at preserving the number type when
encoding as JSON. Therefore some numbers that JSON::XS would dump as integers
are dumped as floats by Cpanel::JSON::XS. This behavior is considered to be acceptable.

Note that in the test descriptons IV means integer and NV means float.