Issue #3937: consistently pass scalars as strings to JS
It is kind of accidental whether numbers are passed as numbers or
as strings to the JavaScript world. Internal changes can lead to
user facing bugs. So let's try to be more predictable and pass scalars
as strings. This requires another tree walker which fiddles with the
data that is to be serialized.

Boolean values are preserved.