Issue #4004: declare the behavior of the bitwise operators
Bitwise operators are rarely used in OTOBO. In all cases the operands
are meant to be taken as numbers. There are no bitwise operations on strings
in OTOBO. So it is safe to activate the feature 'bitwise'.

Adding "use feature 'bitwise';" is mostly a hint for future development. It indicates
that is is safe to activate 'use v5.28;' which implicitly activates the feature 'bitwise'.

"use experimental 'bitwise';" is only added for the sake of the rare cases where
OTOBO actually runs with Perl 5.24 or Perl 5.26.