Issue #4153: make the two SetPassword() methods more similar
These two methods are in Kernel::System::User and in Kernel::System::CustomerUser::DB.
They basically do the same hashification of the passed in password but have superficial diff.
Saintainance is easier when the methods look the same.

A significant difference is that Kernel::System::CustomerUser::DB::SetPassword() sets the
utf8 flag on the hash value. It is believed that this actually makes no difference
as the hashes consists of only low byte iso8859 characters.

Also improve some of the code comments.