Issue #4373: turn the interface modules into Plack apps
Inherit from Plack::Component.
Put the code that was run in new() into the method Response().
Rename the method Response() to call().

Eliminate the custom new(), use the new() provided by Plack::Component.

Explicitly pass the attributes that were formerly in the object.
There were not many object attributes. Basically only 'Debug' and 'TransportObject'
have to be passed explicitly.

Try to make the interface modules more similar to each other, e.g. by having
the exact same comments.

Rename Content() to _Content() in order to indicate that this method is
used only internally.

In Kernel::System::Web::App, call the interface modules like any other Plack app.

The funny way in which the PSGI Env is made available still persists.
This means that the shiny new apps simply ignore the  $Env that is passed to call().

prepare_app() is not yet used.