Sync cpanfile.docker.kerberos with cpanfile.docker
cpanfile.docker is generated by bin/otobo.CheckModules.pl. This works.
cpanfile.docker.kerberos should be the same as cpanfile.docker except that
there is one additional module. The current workflow is that this must be done manully.

The manual update was actually forgotten about and only noticed when Test::Strict was added
is cpanfile.docker. This broke the GitHub action DockerImageBuilderDevel.