Issue #5077: start with clean_slate() and copy_otobo_update()
This change is for building up /opt/otobo from a clean slate during upgrade.
For that the old content of /opt/otobo is copied to a backup directory.
The backup directory is a time stamped subdirectory of the new volume /opt/otobo_backup.
Specific files and directories are copied from the backup to the new installation.
The directory var/article gets special treatment is the content may be large.