Issue #1577: first stab at the new test script OpenLDAP.t
The module Test2::Require::OTOBO::OpenLDAP checks the availability of the LDAP server.

The port of the LDAP has been set to 1389. The default of Net::LDAP is 389.

There is a check whether authentication with the initially provided test users work.

Fixtures are imported from a LDIF file into a subtree that has a distinctive name with a random id.
Limit the searches in the test that follow to that sandbox.

The initial tests are simple authentication checks for the imported testuser.