Issue #1447: Consider existence of feature as package.
Until now, the feature to remove the process information for both agent
and customer interface only existed as PR for core. But recently, the
customer part of the feature was created as a package for OTOBO 11.0.

This needs to be taken into account in the migration script as well as
in the package backend for checking and uninstalling the package during
migrations.