Issue #1447: Added customer setting handling.
The dynamic field screen setting for CustomerTicketZoom needs to be
adapted as well in the migration process.

Additionally, the abortion in case of the installed package needs to be
placed behind the agent-side handling and before the customer-side
handling, as the package only affects the customer side of the feature.