Issue #4705: Lenses on CustomerUser and CustomerCompany fields.
To make lenses with reference fields to customer company and customer
users working, in several places it is needed to pass the object name
instead of the object id.

Additionally, the dynamic field backend ValueSet checks if only one of
ObjectID or ObjectName is passed. Because of that, it is necessary to
tidy the other attribute from %Param in this case.

Co-authored-by: Sven <svo@otobo.org>