Update the CPAN snapshot
MIME-Charset now has Encode::HanExtra as a requirement. This has no
repercussionion in which versions end up in the Docker image as
the requirements listing is only informational.

But the change highlights that optional requirements that exist suddenly
become listed as requirements. Apparently the file content depends
on the order in which distributions are installed. This is the reason for many spurious
changes in cpanfile.docker.snapshot. It is not obvious how to improve the situation.